# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "A-Share Stock Backtesting & Quant Platform API",
    "version": "1.0.0",
    "description": "A-share 量化平台后端 REST API。\n\n主要业务领域：\n- **认证 & 用户**：登录/注册/JWT 刷新\n- **回测 & 策略**：单股回测、组合级策略、参数优化\n- **因子**：因子库管理、IC 报告、相关性矩阵\n- **模拟交易**：持仓、订单、风控守卫\n- **量化引擎**：StrategyEngine / SignalEngine / BacktestEngine 等 5 大 facade\n- **数据 & 行情**：股票元信息、K 线、实时行情、龙虎榜、北向资金\n- **AI**：TradingAgents 深度分析、KOL 观点、公告 NLP、市场情绪\n- **推送 & 告警**：飞书 / 邮件 / 微信 / 短信 4 channel\n- **运维 & 监控**：env 校验、CI 状态、健康度检查\n\n所有 endpoint 默认返回 `{ success: boolean, data?: any, message?: string }` envelope。\n需要鉴权的 endpoint 需在 Header 携带 `Authorization: Bearer <JWT>`。",
    "contact": {
      "name": "QuantX A-Share Alpha Refactor"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development (本地)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在 Authorization Header 携带 `Bearer <JWT>`。Token 通过 `/api/auth/login` 获取。"
      }
    },
    "schemas": {
...
      "Journal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "review_date": {
            "type": "string",
            "format": "date"
          },
          "mood": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "ai_summary": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/journals": {
      "get": {
        "tags": [
          "日记 Journals"
        ],
        "summary": "获取当前用户的复盘日记列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "401": {
            "description": "未授权"
          }
        }
      },
      "post": {
        "tags": [
          "日记 Journals"
        ],
        "summary": "创建复盘日记",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "review_date"
                ],
                "properties": {
                  "review_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "mood": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/journals/{date}": {
      "get": {
        "tags": [
          "日记 Journals"
        ],
        "summary": "获取指定日期的复盘日记详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "日记不存在"
          }
        }
      },
      "put": {
        "tags": [
          "日记 Journals"
        ],
        "summary": "更新复盘日记",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mood": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "日记不存在"
          }
        }
      },
      "delete": {
        "tags": [
          "日记 Journals"
        ],
        "summary": "删除复盘日记",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "日记不存在"
          }
        }
      }
    },
    "/api/journals/{date}/notes": {
      "post": {
        "tags": [
          "日记 Journals"
        ],
        "summary": "US-017 追加用户手记（不修改 AI 生成的字段）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "note"
                ],
                "properties": {
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "日记不存在"
          }
        }
      }
    },
...
    "/api/announcements": {
      "get": {
        "tags": [
          "公告 Announcements"
        ],
        "summary": "某只股票最近 N 天的公告 NLP 抽取列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "stock_code",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "6 位股票代码"
          },
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "default": 30,
              "maximum": 365
            },
            "description": "回看天数"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 200,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/announcements/by-date": {
      "get": {
        "tags": [
          "公告 Announcements"
        ],
        "summary": "某交易日全市场公告列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": true,
            "description": "YYYY-MM-DD"
          },
          {
            "in": "query",
            "name": "sentiment",
            "schema": {
              "type": "string",
              "enum": [
                "正面",
                "中性",
                "负面"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 200,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/announcements/sync": {
      "post": {
        "tags": [
          "公告 Announcements"
        ],
        "summary": "手动触发某日同步 (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "YYYY-MM-DD (默认今日)"
                  },
                  "symbol": {
                    "type": "string",
                    "default": "全部",
                    "description": "全部/重大事项/..."
                  },
                  "extract_with_ai": {
                    "type": "boolean",
                    "default": false
                  },
                  "dry_run": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
...
  }
}